refactor(sync): extract reminder-twin reconciliation into apply/reminders.rs
